+++ authentication/services.py
@@ -1,12 +1,87 @@
-from typing import Tuple
+from typing import Tuple, Dict, Any
 
-from django.db import transaction
 from django.core.management.utils import get_random_secret_key
+from django.core.exceptions import ValidationError
+from django.conf import settings
+from django.http import HttpResponse
+from django.db import transaction
 
 from three60.utils import get_now
 
-from .models import User
+from authentication.models import User
+from authentication.models import User
 
+import jwt, datetime, requests
+
+GOOGLE_ID_TOKEN_INFO_URL = 'https://www.googleapis.com/oauth2/v3/tokeninfo'
+GOOGLE_ACCESS_TOKEN_OBTAIN_URL = 'https://oauth2.googleapis.com/token'
+GOOGLE_USER_INFO_URL = 'https://www.googleapis.com/oauth2/v3/userinfo'
+
+
+def jwt_login(*, response: HttpResponse, user: User) -> HttpResponse:
+    payload = {
+        "id": user.id,
+        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=24),
+        "iat": datetime.datetime.utcnow(),
+    }
+
+    token = jwt.encode(payload, "secret", algorithm="HS256") # generates access token for login
+
+    response.set_cookie(key="jwt", value=token, httponly=True) # creates cookies for user session
+
+
+    return response
+
+
+def google_validate_id_token(*, id_token: str) -> bool:
+    # Reference: https://developers.google.com/identity/sign-in/web/backend-auth#verify-the-integrity-of-the-id-token
+    response = requests.get(
+        GOOGLE_ID_TOKEN_INFO_URL,
+        params={'id_token': id_token}
+    )
+
+    if not response.ok:
+        raise ValidationError('id_token is invalid.')
+
+    audience = response.json()['aud']
+
+    if audience != settings.GOOGLE_OAUTH2_CLIENT_ID:
+        raise ValidationError('Invalid audience.')
+
+    return True
+
+
+def google_get_access_token(*, code: str, redirect_uri: str) -> str:
+    # Reference: https://developers.google.com/identity/protocols/oauth2/web-server#obtainingaccesstokens
+    data = {
+        'code': code,
+        'client_id': settings.GMAIL_API_CLIENT_ID,
+        'client_secret': settings.GMAIL_API_CLIENT_SECRET,
+        'redirect_uri': redirect_uri,
+        'grant_type': 'authorization_code'
+    }
+
+    response = requests.post(GOOGLE_ACCESS_TOKEN_OBTAIN_URL, data=data)
+
+    if not response.ok:
+        raise ValidationError('Failed to obtain access token from Google.')
+
+    access_token = response.json()['access_token']
+
+    return access_token
+
+
+def google_get_user_info(*, access_token: str) -> Dict[str, Any]:
+    # Reference: https://developers.google.com/identity/protocols/oauth2/web-server#callinganapi
+    response = requests.get(
+        GOOGLE_USER_INFO_URL,
+        params={'access_token': access_token}
+    )
+
+    if not response.ok:
+        raise ValidationError('Failed to obtain user info from Google.')
+
+    return response.json()
 
 def user_create(email, password=None, **extra_fields) -> User:
     extra_fields = {
