+++ authentication/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import redirect
 from django.contrib.sites.shortcuts import get_current_site
 from django.conf import settings
 from django.urls import reverse
@@ -6,13 +6,19 @@ from django.urls import reverse
 from rest_framework.generics import GenericAPIView
 from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework import status, serializers
+from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import RefreshToken
-from three60.utils import send_mail
-from api.mixins import ApiAuthMixin
-from .serializers import UserSerializer, RegisterSerializer
-from .models import User
 
+from three60.mixins import ApiErrorsMixin, ApiAuthMixin, PublicApiMixin
+from three60.utils import send_mail
+
+from authentication.services import user_get_or_create, google_get_access_token, google_validate_id_token, google_get_user_info, jwt_login
+from authentication.selectors import get_user
+from authentication.serializers import UserSerializer, RegisterSerializer
+from authentication.models import User
+
+from urllib.parse import urlencode
 
 import jwt,datetime
 
@@ -31,6 +37,7 @@ class RegisterView(GenericAPIView, ApiAuthMixin):
 
         token = RefreshToken.for_user(user).access_token
         current_site = get_current_site(request).domain
+        print(current_site)
         relative_url = reverse('verify-email')
         absolute_url  = "http://" + current_site + relative_url + "?token=" + str(token)
         body = f"Hi {user.username}, verify your email with  this link \n {absolute_url}"
@@ -46,8 +53,11 @@ class RegisterView(GenericAPIView, ApiAuthMixin):
 
 
 class VerifyEmailView(GenericAPIView):
+    swagger_schema = None
     def get(self, request):
         token =request.GET.get('token')
+
+
 
         try:
             payload = jwt.decode(token, settings.SECRET_KEY, algorithms=['HS256'])
@@ -56,7 +66,7 @@ class VerifyEmailView(GenericAPIView):
             if not user.is_verified:
                 user.is_verified = True
                 user.save()
-            return Response({'success':'user successfully created'}, status=status.HTTP_202_ACCEPTED)
+            return Response(redirect(f'{settings.BASE_FRONTEND_URL}\login'))
 
         except jwt.ExpiredSignatureError as identifier:
             return Response(
@@ -96,4 +106,77 @@ class LoginView(GenericAPIView):
         # token = user.tokens()
         # response.set_cookie(key="jwt", value=token.access_token, httponly=True) # creates cookies for user session
         response.data = {"tokens": user.tokens(), "data":serializer.data}
+        return response
+
+
+class GoogleLoginApi(PublicApiMixin, ApiErrorsMixin, APIView):
+    class InputSerializer(serializers.Serializer):
+        code = serializers.CharField(required=False)
+        error = serializers.CharField(required=False)
+
+    def get(self, request, *args, **kwargs):
+        input_serializer = self.InputSerializer(data=request.GET)
+        input_serializer.is_valid(raise_exception=True)
+
+        validated_data = input_serializer.validated_data
+
+        code = validated_data.get('code')
+        error = validated_data.get('error')
+
+        login_url = f'{settings.BASE_FRONTEND_URL}/login'
+
+        if error or not code:
+            params = urlencode({'error': error})
+            return redirect(f'{login_url}?{params}')
+
+        domain = get_current_site(request).domain
+        api_uri = reverse('google-login')
+        redirect_uri = f'{domain}{api_uri}'
+
+        access_token = google_get_access_token(code=code, redirect_uri=redirect_uri)
+
+        user_data = google_get_user_info(access_token=access_token)
+
+        profile_data = {
+            'email': user_data['email'],
+            'first_name': user_data.get('givenName', ''),
+            'last_name': user_data.get('familyName', ''),
+        }
+
+        # We use get-or-create logic here for the sake of the example.
+        # We don't have a sign-up flow.
+        user, _ = user_get_or_create(**profile_data)
+
+        response = redirect(settings.BASE_FRONTEND_URL)
+        response = jwt_login(response=response, user=user)
+
+        return response
+
+
+class UserApi(ApiAuthMixin, ApiErrorsMixin, APIView):
+    swagger_schema = None
+    def get(self, request):
+        return Response(get_user(user=request.user))
+
+
+class UserInitApi(PublicApiMixin, ApiErrorsMixin, APIView):
+    swagger_schema = None
+    class InputSerializer(serializers.Serializer):
+        email = serializers.EmailField()
+        first_name = serializers.CharField(required=False, default='')
+        last_name = serializers.CharField(required=False, default='')
+    def post(self, request, *args, **kwargs):
+        id_token = request.headers.get('Authorization')
+        google_validate_id_token(id_token=id_token)
+
+        serializer = self.InputSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+
+        # We use get-or-create logic here for the sake of the example.
+        # We don't have a sign-up flow.
+        user, _ = user_get_or_create(**serializer.validated_data)
+
+        response = Response(data=get_user(user=user))
+        response = jwt_login(response=response, user=user)
+
         return response
