--- backend/three60/social/views.py
@@ -1,12 +1,0 @@
-from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter
-from allauth.socialaccount.providers.oauth2.client import OAuth2Client
-from dj_rest_auth.registration.views import SocialLoginView
-from django.conf import settings
-
-
-# Create your views here.
-class GoogleLogin(SocialLoginView):
-    authentication_classes = [] # disable authentication
-    adapter_class = GoogleOAuth2Adapter
-    callback_url = "http://localhost:3000"
-    client_class = OAuth2Client
