+++ three60/.env
@@ -3,6 +3,6 @@ OAUTH_CLIENT_SECRET = "uJVdltzQNrHHESgocQZpfmppnSNGzravQixnTbZgJJO4olcSBj98BT9H1
 OAUTH_CLIENT_ID = "S5erwtmUIaUsSN9burAPKpiGMBLK3nNRRw38bJgW"
 USER = 'three60'
 PASSWORD = 'oyeladunkola'
-GMAIL_API_CLIENT_ID = '335239736987-10c3pogthhojtu2t720kc3lu4e9k6tba.apps.googleusercontent.com'
-GMAIL_API_CLIENT_SECRET = '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'
+CLIENT_IID = '335239736987-10c3pogthhojtu2t720kc3lu4e9k6tba.apps.googleusercontent.com'
+CLIENT_SECRET = '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'
 GMAIL_API_REFRESH_TOKEN = '1//048oXaMuc_htPCgYIARAAGAQSNwF-L9IrxvE3ZI9c3HZKxqOrlDakfDfANATnvT7l-_NZWaHTqs45m7rA9ab4LUV6dHHsuwjAXyo'
