+++ three60/urls.py
@@ -32,6 +32,5 @@ schema_view = get_schema_view(
 urlpatterns = [
     path("admin/", admin.site.urls),
     path('', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
-    path('social/', include('social.urls')),
-    path('api/auth/', include('authentication.urls'))
+    path('auth/', include('authentication.urls'), name = 'auth'),
 ]
